explained all the commands in the help function
also cls isn't working while it's asking the folder name. added it but somehow it isn't working.

added a function to clear the screen so we can just call the functions whenever the need arises.

and the splitting thing isn't working.. idk why.. it just goes on a loop when it tries to ask how many files would they take per folder.