fixed a bug, improved functions.
like an idiot, i assumed that python reads the files and lists them
alphabetically. which wasn't true when i tried to split my fire sime. my
code's logic was right. but the inital file structure that i was getting
was not right. so i fixed that with a sort function.

also improved some stuff.. like printing out stuff, and also lists.